Move adjustFigAspect function from validate.py to utils.py
