+++ src/utils.py
@@ -85,6 +85,26 @@ def get_dict_from_args(args):
     return config
 
 # %%
+''' Functions below are used in validate.py '''
+# %%
+def adjustFigAspect(fig,aspect=1):
+    ''' Adjust the subplot parameters so that the figure has the correct aspect ratio.'''
+    xsize,ysize = fig.get_size_inches()
+    minsize = min(xsize,ysize)
+    xlim = .4*minsize/xsize
+    ylim = .4*minsize/ysize
+    if aspect < 1:
+        xlim *= aspect
+    else:
+        ylim /= aspect
+    fig.subplots_adjust(left=.5-xlim,
+                        right=.5+xlim,
+                        bottom=.5-ylim,
+                        top=.5+ylim)
+
+# %%
+
+# %%
 if __name__=='__main__':
     parser = argparse.ArgumentParser(description='Training Hyperparameters')
     args = parse_arguments(parser)
